Added handling of trailing comments after a chunk
This is relevant e.g. at the end of a file where comments were being
ignored before. Now they are saved as a member of the Chunk class.

Added support for these in the python pretty string and lua printers.

Adjusted one test to confirm to the new printer output and wrote two new
ones.

The implementation is possibly(?) hacky and once the lua output is done,
much more tests are needed there.