Moved most comment functionality from chunks to blocks.
Nested blocks, like if statements, can have comments, esp. trailing
comments too. Hence blocks need to support them and not throw them away
as before.
As all chunks have a block as the next element in the tree, this outer
block can, and should, just take over (most) of the comments that were
previously assigned to the chunk.