Implemented parenthesis around binary operators.
It is very weird that they weren't implemented before, considering that
the order of e.g. arithmetic operations depends heavily on parenthesis
being part of those operations or not.

My implementation seems to work for all cases I tested it, but I'm also
fairly certain that there is a much better way to implement it,
considering that everything I did feels like a giant hack. As usual in
this repo.